commit 4

+++ homework1/src/main/java/com/ssk/handler/CustomerHandler.java
@@ -3,29 +3,33 @@ package com.ssk.handler;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.server.ServerRequest;
 import org.springframework.web.reactive.function.server.ServerResponse;
 
 import com.ssk.domain.Customer;
+import com.ssk.exception.BadRequestException;
 import com.ssk.service.CustomerService;
 
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-//@ApiResponses({
-//	   @ApiResponse(responseCode = "200", description = "OK !!"),
-//       @ApiResponse(responseCode = "400", description = "BAD REQUEST !!"),
-//       @ApiResponse(responseCode = "404", description = "NOT FOUND !!"),
-//       @ApiResponse(responseCode = "500", description = "INTERNAL SERVER ERROR !!")
-//})
+
 @Component
 public class CustomerHandler {
 
 	@Autowired
-	private CustomerService service;
+	private final CustomerService service;
 
+	public CustomerHandler(CustomerService service) {
+		super();
+		this.service = service;
+	}
+
+	public boolean isNumeric (String id) {
+		return id.matches("[+-]?\\d*(\\.\\d+)?");
+	}
+	
 	public Mono<ServerResponse> listCustomer(ServerRequest request) {
 		Flux<Customer> customers = service.findAll();
 		return ServerResponse.ok().contentType(APPLICATION_JSON).body(customers, Customer.class);
@@ -33,25 +37,30 @@ public class CustomerHandler {
 
 	public Mono<ServerResponse> getCustomer(ServerRequest request) {
 		String id = request.pathVariable("id");
-		Mono<Customer> customer = service.findById(id);
+		if(!isNumeric(id)) throw new BadRequestException();
+		Mono<Customer> customer = service.findById(id);	
 		return ServerResponse.ok().contentType(APPLICATION_JSON).body(customer, Customer.class);
+     
 	}
-
+	
 	public Mono<ServerResponse> createCustomer(ServerRequest request) {
-		Mono<Customer> customer = request.bodyToMono(Customer.class).flatMap(c -> service.save(c));
+		Mono<Customer> customer = request.bodyToMono(Customer.class).flatMap(c -> service.save(c))
+				.onErrorResume(e -> Mono.error(new BadRequestException(e)));  
 		return ServerResponse.ok().contentType(APPLICATION_JSON).body(customer, Customer.class);
 	}
 
 	public Mono<ServerResponse> deleteCustomer(ServerRequest request) {
 		String id = request.pathVariable("id");
+		if(!isNumeric(id)) throw new BadRequestException();
 		return ServerResponse.ok().contentType(APPLICATION_JSON).body(service.delete(id), Customer.class);
 	}
 
 	public Mono<ServerResponse> updateCustomer(ServerRequest request) {
 		String id = request.pathVariable("id");
-		Mono<Customer> customer = request.bodyToMono(Customer.class);
-		return customer
-				.flatMap(c -> ServerResponse.status(HttpStatus.CREATED).body(service.update(c, id), Customer.class));
+		if(!isNumeric(id)) throw new BadRequestException();
+		Mono<Customer> customer = request.bodyToMono(Customer.class).flatMap(c -> service.update(c,id))
+				.onErrorResume(e -> Mono.error(new BadRequestException(e)));  
+	    return ServerResponse.ok().contentType(APPLICATION_JSON).body(customer, Customer.class);
 	}
 
 }
